feat: ability to pass props to a custom react component
Component properties are defined in Python, serialized to a dictionary,
and sent to the DOM via `data-props` HTML attribute.

`makeReactInput` function receives those props, parses and sends
them further to the React component. This is possible thanks
to a second generic parameter which defines the shape of the props.

To showcase this new feature, the `demo-pkg` code is updated
accordingly.